add pipeline, py.typed marker and sort out imports
