+++ .github/workflows/run-tests-and-mypy.yml
@@ -1,0 +1,41 @@
+name: Run tests and mypy
+
+on:
+  push:
+      branches: [ main ]
+  pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
+
+
+jobs:
+
+  unit_tests:
+    if: |
+      github.event_name == 'push' ||
+      (github.event_name == 'pull_request' && github.event.pull_request.draft == false)
+    name: Unit testing
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.8]
+
+    steps:
+
+        - name: Checkout main code and submodules
+          uses: actions/checkout@v2
+
+        - name: Set up Python ${{ matrix.python-version }}
+          uses: actions/setup-python@v2
+          with:
+            python-version: ${{ matrix.python-version }}
+
+        - name: Install main dependencies
+          run: |
+            python -m pip install --upgrade pip
+            pip install . 
+        - name: Perform unit tests
+          run: |
+            pytest -vvv --cov pyqtorch tests
+        - name: Mypy checks
+          run: |
+            mypy pyqtorch
