+++ pyqtorch/__init__.py
@@ -11,3 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from .core import measurement, utils
+from .core.batched_operation import *
+from .core.circuit import *
+from .core.measurement import measure_openfermion
+from .core.operation import *
