+++ setup.py
@@ -40,6 +40,7 @@ setup(
     install_requires=REQUIRED,
     extras_require=EXTRAS,
     include_package_data=True,
+    package_data={"pyqtorch": ["py.typed"]},
     license="Apache 2.0",
     classifiers=[
         "License :: Other/Proprietary License",
